core: Move runtime error logic to submodule
