+++ src/core/error_handling/RuntimeErrorCollector.cpp
@@ -16,7 +16,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "RuntimeErrorCollector.hpp"
+
+#include "error_handling/RuntimeErrorCollector.hpp"
 
 #include <utils/mpi/gather_buffer.hpp>
