+++ src/core/error_handling/RuntimeErrorCollector.hpp
@@ -17,10 +17,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef ERROR_HANDLING_RUNTIMEERRORCOLLECTOR_HPP
-#define ERROR_HANDLING_RUNTIMEERRORCOLLECTOR_HPP
+#ifndef ESPRESSO_SRC_CORE_ERROR_HANDLING_RUNTIME_ERROR_COLLECTOR_HPP
+#define ESPRESSO_SRC_CORE_ERROR_HANDLING_RUNTIME_ERROR_COLLECTOR_HPP
 
-#include "RuntimeError.hpp"
+#include "error_handling/RuntimeError.hpp"
 
 #include <boost/mpi/communicator.hpp>
