+++ src/core/errorhandling.cpp
@@ -25,7 +25,7 @@
 #include "errorhandling.hpp"
 
 #include "MpiCallbacks.hpp"
-#include "RuntimeErrorCollector.hpp"
+#include "error_handling/RuntimeErrorCollector.hpp"
 
 #include <boost/mpi/collectives.hpp>
 #include <boost/mpi/communicator.hpp>
