+++ src/core/errorhandling.hpp
@@ -20,17 +20,16 @@
  */
 /** \file
  *  This file contains the errorhandling code for severe errors, like
- *  a broken bond or illegal parameter combinations. See section
- *  "Errorhandling for developers" for details on the error format and
- *  how to use this.
+ *  a broken bond or illegal parameter combinations.
  */
-#ifndef ERRORHANDLING_HPP
-#define ERRORHANDLING_HPP
+
+#ifndef ESPRESSO_SRC_CORE_ERROR_HANDLING_HPP
+#define ESPRESSO_SRC_CORE_ERROR_HANDLING_HPP
 
 #include "config.hpp"
 
-#include "RuntimeError.hpp"
-#include "RuntimeErrorStream.hpp"
+#include "error_handling/RuntimeError.hpp"
+#include "error_handling/RuntimeErrorStream.hpp"
 
 #include <string>
 #include <vector>
