+++ src/script_interface/ParallelExceptionHandler.cpp
@@ -22,7 +22,7 @@
 #include "Exception.hpp"
 
 #include "core/MpiCallbacks.hpp"
-#include "core/RuntimeError.hpp"
+#include "core/error_handling/RuntimeError.hpp"
 #include "core/communication.hpp"
 #include "core/errorhandling.hpp"
